requirements.txt
# MemOS v2.0 Dependencies

# LLM & LangChain
langchain-openai>=0.2.0
openai>=1.0.0

# Database
supabase>=2.0.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0

# Document Processing (Optional)
PyPDF2>=3.0.0
python-docx>=1.1.0
